Reworked parameter passing to module in lxc_container
Previously, each possible argument for the proxmox module called
by the lxc_container role was a separate top-level role variable.
This made it impossible to omit unwanted parameters (such as the
password for the root user) from the creation API call. As the PVE
API treats undefined parameters differently from empty parameters,
thic meant that certain actions were not possible through the role.

For example, if the user did not want to set a root password, they
could not set the lxccreate_password role to '!' or '', as the PVE
API would not accept such a value. Omitting the password entirely
is allowed however and will result in the root password being
disabled - the desired behavior in this case.

To account for situations like these, the role variables have been
reworked significantly. All non-essential module arguments are now
collected in a single dict 'lxccreate_args' and then passed to
the proxmox module along with a few other things like API information.
The goal of this change is to allow for more flexibility, both now
and in the future: It allows useres to fully control which parameters
should be set during container creation.

Unfortunatly, this change is utterly incompaible with the old method,
which means that a major version bump is required.