pbs_client: remove support for ubuntu
the upstream proxmox repository no longer mentions Ubuntu as supported,
and both 20.04 and 22.04 have required workarounds in the past.
While 24.04 currently seems to work with the bookworm repository,
this is not guaranteed and ultimately not something i want to support
going forward.

This patch thus removes support for Ubuntu nodes in the pbs_client role.